# backend/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "Whitfield WMS API",
  "version": "1.0.0"
 },
 "paths": {
...
  "/v1/sellers": {
   "post": {
    "tags": [
     "Sellers"
    ],
    "summary": "Create Seller",
    "description": "Create a seller (admin only).\n\nArgs:\n    request (SellerCreate): Seller payload.\n    auth (dict): Authenticated user claims.\n\nReturns:\n    SellerResponse: Created seller.",
    "operationId": "create_seller_v1_sellers_post",
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/SellerCreate"
       }
      }
     }
    },
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/SellerResponse"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "get": {
    "tags": [
     "Sellers"
    ],
    "summary": "List Sellers",
    "description": "List sellers.\n\nArgs:\n    is_active (Optional[bool]): Active filter.\n    auth (dict): Authenticated user claims.\n\nReturns:\n    list[SellerResponse]: Sellers.",
    "operationId": "list_sellers_v1_sellers_get",
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "parameters": [
     {
      "name": "is_active",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "boolean"
        },
        {
         "type": "null"
        }
       ],
       "title": "Is Active"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "array",
         "items": {
          "$ref": "#/components/schemas/SellerResponse"
         },
         "title": "Response List Sellers V1 Sellers Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/v1/sellers/{seller_id}": {
   "get": {
    "tags": [
     "Sellers"
    ],
    "summary": "Get Seller",
    "description": "Fetch a seller by id.\n\nArgs:\n    seller_id (str): Seller id.\n    auth (dict): Authenticated user claims.\n\nReturns:\n    SellerResponse: Seller.",
    "operationId": "get_seller_v1_sellers__seller_id__get",
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "parameters": [
     {
      "name": "seller_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Seller Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/SellerResponse"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "patch": {
    "tags": [
     "Sellers"
    ],
    "summary": "Update Seller",
    "description": "Update a seller (admin only).\n\nArgs:\n    seller_id (str): Seller id.\n    request (SellerUpdate): Update payload.\n    auth (dict): Authenticated user claims.\n\nReturns:\n    SellerResponse: Updated seller.",
    "operationId": "update_seller_v1_sellers__seller_id__patch",
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "parameters": [
     {
      "name": "seller_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Seller Id"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/SellerUpdate"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/SellerResponse"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/v1/products": {
   "post": {
    "tags": [
     "Products"
    ],
    "summary": "Create Product",
    "description": "Create a product.\n\nArgs:\n    request (ProductCreate): Product payload.\n    auth (dict): Authenticated user claims.\n\nReturns:\n    ProductResponse: Created product.",
    "operationId": "create_product_v1_products_post",
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/ProductCreate"
       }
      }
     }
    },
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ProductResponse"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "get": {
    "tags": [
     "Products"
    ],
    "summary": "List Products",
    "description": "List products.\n\nArgs:\n    seller_id (Optional[str]): Seller filter.\n    auth (dict): Authenticated user claims.\n\nReturns:\n    list[ProductResponse]: Products.",
    "operationId": "list_products_v1_products_get",
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "parameters": [
     {
      "name": "seller_id",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Seller Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "array",
         "items": {
          "$ref": "#/components/schemas/ProductResponse"
         },
         "title": "Response List Products V1 Products Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/v1/products/upc/{upc}": {
   "get": {
    "tags": [
     "Products"
    ],
    "summary": "Get Product By Upc",
    "description": "Fetch a product by UPC barcode.\n\nArgs:\n    upc (str): UPC barcode.\n    auth (dict): Authenticated user claims.\n\nReturns:\n    ProductResponse: Product.",
    "operationId": "get_product_by_upc_v1_products_upc__upc__get",
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "parameters": [
     {
      "name": "upc",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Upc"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ProductResponse"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/v1/products/{product_id}": {
   "get": {
    "tags": [
     "Products"
    ],
    "summary": "Get Product",
    "description": "Fetch a product by id.\n\nArgs:\n    product_id (str): Product id.\n    auth (dict): Authenticated user claims.\n\nReturns:\n    ProductResponse: Product.",
    "operationId": "get_product_v1_products__product_id__get",
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "parameters": [
     {
      "name": "product_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Product Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ProductResponse"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "patch": {
    "tags": [
     "Products"
    ],
    "summary": "Update Product",
    "description": "Update a product.\n\nArgs:\n    product_id (str): Product id.\n    request (ProductUpdate): Update payload.\n    auth (dict): Authenticated user claims.\n\nReturns:\n    ProductResponse: Updated product.",
    "operationId": "update_product_v1_products__product_id__patch",
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "parameters": [
     {
      "name": "product_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Product Id"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/ProductUpdate"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ProductResponse"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
  "/v1/bin-locations/product/{product_id}": {
   "get": {
    "tags": [
     "Bin Locations"
    ],
    "summary": "Find Bin For Product",
    "description": "Find the bin where a product is stored.\n\nArgs:\n    product_id (str): Product id.\n    warehouse_id (str): Warehouse id.\n    auth (dict): Authenticated user claims.\n\nReturns:\n    BinLocationResponse: Bin location.",
    "operationId": "find_bin_for_product_v1_bin_locations_product__product_id__get",
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "parameters": [
     {
      "name": "product_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Product Id"
      }
     },
     {
      "name": "warehouse_id",
      "in": "query",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Warehouse Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/BinLocationResponse"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
...
   "BillingRatesCreate": {
    "properties": {
     "storage_per_unit_per_day": {
      "type": "number",
      "title": "Storage Per Unit Per Day",
      "default": 0.05
     },
     "fulfillment_per_order": {
      "type": "number",
      "title": "Fulfillment Per Order",
      "default": 3.5
     },
     "receiving_per_unit": {
      "type": "number",
      "title": "Receiving Per Unit",
      "default": 0.25
     }
    },
    "type": "object",
    "title": "BillingRatesCreate",
    "description": "Per-unit billing rates."
   },
   "BillingRatesResponse": {
    "properties": {
     "storage_per_unit_per_day": {
      "type": "number",
      "title": "Storage Per Unit Per Day"
     },
     "fulfillment_per_order": {
      "type": "number",
      "title": "Fulfillment Per Order"
     },
     "receiving_per_unit": {
      "type": "number",
      "title": "Receiving Per Unit"
     }
    },
    "type": "object",
    "required": [
     "storage_per_unit_per_day",
     "fulfillment_per_order",
     "receiving_per_unit"
    ],
    "title": "BillingRatesResponse",
    "description": "Per-unit billing rates."
   },
...
   "BinLocationResponse": {
    "properties": {
     "id": {
      "type": "string",
      "title": "Id"
     },
     "warehouse_id": {
      "type": "string",
      "title": "Warehouse Id"
     },
     "bin_code": {
      "type": "string",
      "title": "Bin Code"
     },
     "aisle": {
      "type": "string",
      "title": "Aisle"
     },
     "row": {
      "type": "string",
      "title": "Row"
     },
     "shelf": {
      "type": "string",
      "title": "Shelf"
     },
     "bin": {
      "type": "string",
      "title": "Bin"
     },
     "product_id": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Product Id"
     },
     "max_capacity": {
      "type": "integer",
      "title": "Max Capacity"
     },
     "current_units": {
      "type": "integer",
      "title": "Current Units"
     },
     "is_occupied": {
      "type": "boolean",
      "title": "Is Occupied"
     }
    },
    "type": "object",
    "required": [
     "id",
     "warehouse_id",
     "bin_code",
     "aisle",
     "row",
     "shelf",
     "bin",
     "max_capacity",
     "current_units",
     "is_occupied"
    ],
    "title": "BinLocationResponse",
    "description": "Bin location representation."
   },
...
   "DimensionsCreate": {
    "properties": {
     "weight_lbs": {
      "type": "number",
      "title": "Weight Lbs",
      "default": 0.0
     },
     "length_in": {
      "type": "number",
      "title": "Length In",
      "default": 0.0
     },
     "width_in": {
      "type": "number",
      "title": "Width In",
      "default": 0.0
     },
     "height_in": {
      "type": "number",
      "title": "Height In",
      "default": 0.0
     }
    },
    "type": "object",
    "title": "DimensionsCreate",
    "description": "Product physical dimensions."
   },
   "DimensionsResponse": {
    "properties": {
     "weight_lbs": {
      "type": "number",
      "title": "Weight Lbs"
     },
     "length_in": {
      "type": "number",
      "title": "Length In"
     },
     "width_in": {
      "type": "number",
      "title": "Width In"
     },
     "height_in": {
      "type": "number",
      "title": "Height In"
     }
    },
    "type": "object",
    "required": [
     "weight_lbs",
     "length_in",
     "width_in",
     "height_in"
    ],
    "title": "DimensionsResponse",
    "description": "Product physical dimensions."
   },
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
...
   "ProductCreate": {
    "properties": {
     "seller_id": {
      "type": "string",
      "title": "Seller Id"
     },
     "upc_barcode": {
      "type": "string",
      "title": "Upc Barcode"
     },
     "sku": {
      "type": "string",
      "title": "Sku"
     },
     "product_name": {
      "type": "string",
      "title": "Product Name"
     },
     "description": {
      "type": "string",
      "title": "Description",
      "default": ""
     },
     "dimensions": {
      "$ref": "#/components/schemas/DimensionsCreate"
     },
     "low_stock_threshold": {
      "type": "integer",
      "title": "Low Stock Threshold",
      "default": 20
     },
     "category": {
      "type": "string",
      "title": "Category",
      "default": ""
     },
     "is_active": {
      "type": "boolean",
      "title": "Is Active",
      "default": true
     }
    },
    "type": "object",
    "required": [
     "seller_id",
     "upc_barcode",
     "sku",
     "product_name"
    ],
    "title": "ProductCreate",
    "description": "Payload to create a product."
   },
   "ProductResponse": {
    "properties": {
     "id": {
      "type": "string",
      "title": "Id"
     },
     "seller_id": {
      "type": "string",
      "title": "Seller Id"
     },
     "upc_barcode": {
      "type": "string",
      "title": "Upc Barcode"
     },
     "sku": {
      "type": "string",
      "title": "Sku"
     },
     "product_name": {
      "type": "string",
      "title": "Product Name"
     },
     "description": {
      "type": "string",
      "title": "Description"
     },
     "dimensions": {
      "$ref": "#/components/schemas/DimensionsResponse"
     },
     "low_stock_threshold": {
      "type": "integer",
      "title": "Low Stock Threshold"
     },
     "category": {
      "type": "string",
      "title": "Category"
     },
     "is_active": {
      "type": "boolean",
      "title": "Is Active"
     }
    },
    "type": "object",
    "required": [
     "id",
     "seller_id",
     "upc_barcode",
     "sku",
     "product_name",
     "description",
     "dimensions",
     "low_stock_threshold",
     "category",
     "is_active"
    ],
    "title": "ProductResponse",
    "description": "Product representation."
   },
   "ProductUpdate": {
    "properties": {
     "sku": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Sku"
     },
     "product_name": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Product Name"
     },
     "description": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Description"
     },
     "dimensions": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/DimensionsCreate"
       },
       {
        "type": "null"
       }
      ]
     },
     "low_stock_threshold": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Low Stock Threshold"
     },
     "category": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Category"
     },
     "is_active": {
      "anyOf": [
       {
        "type": "boolean"
       },
       {
        "type": "null"
       }
      ],
      "title": "Is Active"
     }
    },
    "type": "object",
    "title": "ProductUpdate",
    "description": "Payload to update a product."
   },
...
   "SellerCreate": {
    "properties": {
     "company_name": {
      "type": "string",
      "title": "Company Name"
     },
     "contact_name": {
      "type": "string",
      "title": "Contact Name"
     },
     "email": {
      "type": "string",
      "title": "Email"
     },
     "phone": {
      "type": "string",
      "title": "Phone"
     },
     "billing_rates": {
      "$ref": "#/components/schemas/BillingRatesCreate"
     },
     "portal_password": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Portal Password"
     },
     "low_stock_threshold_default": {
      "type": "integer",
      "title": "Low Stock Threshold Default",
      "default": 20
     },
     "is_active": {
      "type": "boolean",
      "title": "Is Active",
      "default": true
     }
    },
    "type": "object",
    "required": [
     "company_name",
     "contact_name",
     "email",
     "phone"
    ],
    "title": "SellerCreate",
    "description": "Payload to create a seller."
   },
   "SellerResponse": {
    "properties": {
     "id": {
      "type": "string",
      "title": "Id"
     },
     "company_name": {
      "type": "string",
      "title": "Company Name"
     },
     "contact_name": {
      "type": "string",
      "title": "Contact Name"
     },
     "email": {
      "type": "string",
      "title": "Email"
     },
     "phone": {
      "type": "string",
      "title": "Phone"
     },
     "billing_rates": {
      "$ref": "#/components/schemas/BillingRatesResponse"
     },
     "low_stock_threshold_default": {
      "type": "integer",
      "title": "Low Stock Threshold Default"
     },
     "is_active": {
      "type": "boolean",
      "title": "Is Active"
     }
    },
    "type": "object",
    "required": [
     "id",
     "company_name",
     "contact_name",
     "email",
     "phone",
     "billing_rates",
     "low_stock_threshold_default",
     "is_active"
    ],
    "title": "SellerResponse",
    "description": "Seller representation (no portal password hash)."
   },
   "SellerUpdate": {
    "properties": {
     "contact_name": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Contact Name"
     },
     "email": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Email"
     },
     "phone": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Phone"
     },
     "billing_rates": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/BillingRatesCreate"
       },
       {
        "type": "null"
       }
      ]
     },
     "low_stock_threshold_default": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Low Stock Threshold Default"
     },
     "is_active": {
      "anyOf": [
       {
        "type": "boolean"
       },
       {
        "type": "null"
       }
      ],
      "title": "Is Active"
     }
    },
    "type": "object",
    "title": "SellerUpdate",
    "description": "Payload to update a seller."
   },
...
   "ValidationError": {
    "properties": {
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "type": "array",
      "title": "Location"
     },
     "msg": {
      "type": "string",
      "title": "Message"
     },
     "type": {
      "type": "string",
      "title": "Error Type"
     }
    },
    "type": "object",
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError"
   },
...
  },
  "securitySchemes": {
   "OAuth2PasswordBearer": {
    "type": "oauth2",
    "flows": {
     "password": {
      "scopes": {},
      "tokenUrl": "/v1/auth/login"
     }
    }
   }
  }
 }
}